Поправил замечания
