+++ src/view/route-total-view.js
@@ -1,4 +1,5 @@
 import { createElement } from '../render.js';
+import { howManyCities } from '../utils/utils.js';
 
 const calcTotal = (arr) => {
   const initialValue = 0;
@@ -8,14 +9,6 @@ const calcTotal = (arr) => {
   );
 };
 
-const howManyCities = (arr) => {
-  if (arr.length === 1) {
-    return `${arr[0].reachPoint}`;
-  } else if (arr.length === 2) {
-    return `${arr[0].reachPoint} &mdash; ${arr[1].reachPoint}`;
-  }
-  return `${arr[0].reachPoint} &mdash; ${arr.length > 3 ? '...' : arr[1].reachPoint} &mdash; ${arr[arr.length - 1].reachPoint}`;
-};
 
 const createRoute = (arr) => (
   `<section class="trip-main__trip-info  trip-info">
